backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
anthropic==0.51.0
python-dotenv==1.1.0
